check if logs exist

+++ internal/test/e2e/setup.go
@@ -2,6 +2,7 @@ package e2e
 
 import (
 	"fmt"
+	"os"
 	"strconv"
 
 	"github.com/aws/aws-sdk-go/aws/session"
@@ -157,8 +158,10 @@ func (e *E2ESession) downloadRequiredFileInInstance(file string) error {
 }
 
 func (e *E2ESession) uploadLogFilesFromInstance(testName string) {
+	if !e.ifLogsExist() {
+		return
+	}
 	logger.V(1).Info("Uploading log files to s3 bucket")
-
 	testNameFolder := testName + e.instanceId
 	command := fmt.Sprintf("aws s3 cp /home/e2e/%s/logs s3://%s/logs/%s --recursive", e.instanceId, e.storageBucket, testNameFolder)
 
@@ -181,4 +184,10 @@ func (e *E2ESession) downloadRequiredFilesInInstance() error {
 	}
 
 	return nil
+}
+
+func (e *E2ESession) ifLogsExist() bool {
+	logsFolder := "/home/e2e/" + e.instanceId + "/logs"
+	_, err := os.Stat(logsFolder)
+	return !os.IsNotExist(err)
 }
